Version 0.1.0
Importing the project's fonts and creating the global styles of the application.

Importing the fonts that will be used in the project in the 'index.html' file of the application. Creating the global style files for the application, using the 'styles' folder

Creating the base files and setting up the environment for the creation of pages and components.

Creating the base files of the application containing initial configurations. Creating within the 'utils' folder, the 'alerts' folder with the 'index.js' file, which contains functions to trigger custom alerts for the user. Also creating the 'validations' and 'roles' folders to store validations through regular expressions and make the possible user 'roles' available to the rest of the application, respectively. Finally, creating the authentication context that contains the business logic related to user login and creating the application's route files, preparing the structure for the development of components and pages.

Creating the 'Button', 'Logo', and 'Input' components, and building the 'SignIn' and 'SignUp' pages

Creating the initial components of the application, the 'Button', 'Logo' and 'Input' components within the 'components' folder, with each component having its respective 'index.jsx' and 'styles.js' files to contain the structure and styles of the component, respectively. Storing the images used in the project so far in the 'images' folder within the 'public' folder to render the images dynamically. Updating the application's route files by changing the file extension to 'jsx' and modifying the 'auth.routes.jsx' file. Updating the global styles of the application by adding custom alert styles. Modifying the regular expressions for password and price within the 'validations' folder. Finally, creating the pages for the authentication route, the SignIn (Login) page, and the SignUp (Registration) page, and adding these two routes to the 'auth.routes.jsx' file.

Creating the components for the home page and the page containing the details of the dish.

Creating the final pages of the application, 'NewDish' and 'EditDish,' and preparing the application for deployment

Building the components for the latest pages of the application, creating the 'NewDish' and 'EditDish' components for the admin route, responsible for creating a new dish and updating an existing one, respectively, along with the components that make up these pages. Updating existing components and pages to enhance the site's performance and usability. Improving the layout of custom alerts on both smaller and larger screens. Adding all project routes. Finally, preparing the project for deployment by updating the API address in the 'api.js' file.

Solving prodution problems

Preparing the environment to create the documentation for the application on Github.

Fixing the names of the images and directories in the '.github' folder

Creating the README.md file and finishing the first version of application