Creating the base files and setting up the environment for the creation of pages and components.
Creating the base files of the application containing initial configurations. Creating within the 'utils' folder, the 'alerts' folder with the 'index.js' file, which contains functions to trigger custom alerts for the user. Also creating the 'validations' and 'roles' folders to store validations through regular expressions and make the possible user 'roles' available to the rest of the application, respectively. Finally, creating the authentication context that contains the business logic related to user login and creating the application's route files, preparing the structure for the development of components and pages.