Creating the 'Button', 'Logo', and 'Input' components, and building the 'SignIn' and 'SignUp' pages
Creating the initial components of the application, the 'Button', 'Logo' and 'Input' components within the 'components' folder, with each component having its respective 'index.jsx' and 'styles.js' files to contain the structure and styles of the component, respectively. Storing the images used in the project so far in the 'images' folder within the 'public' folder to render the images dynamically. Updating the application's route files by changing the file extension to 'jsx' and modifying the 'auth.routes.jsx' file. Updating the global styles of the application by adding custom alert styles. Modifying the regular expressions for password and price within the 'validations' folder. Finally, creating the pages for the authentication route, the SignIn (Login) page, and the SignUp (Registration) page, and adding these two routes to the 'auth.routes.jsx' file.