Creating the final pages of the application, 'NewDish' and 'EditDish,' and preparing the application for deployment
Building the components for the latest pages of the application, creating the 'NewDish' and 'EditDish' components for the admin route, responsible for creating a new dish and updating an existing one, respectively, along with the components that make up these pages. Updating existing components and pages to enhance the site's performance and usability. Improving the layout of custom alerts on both smaller and larger screens. Adding all project routes. Finally, preparing the project for deployment by updating the API address in the 'api.js' file.